+++ app_v2/auth/auth_handler.py
@@ -5,8 +5,9 @@ from fastapi import HTTPException
 from passlib.context import CryptContext
 
 from config import settings
-from schemas import AccessToken
+from schemas.auth import AccessToken
 
+_hasher = CryptContext(schemes=["bcrypt"])
 
 def encode_jwt(username: str) -> AccessToken:
     headers = {
@@ -32,17 +33,15 @@ def decode_jwt(token: str) -> dict:
         decoded_token = jwt.decode(token, issuer=settings.token_issuer, key=settings.jwt_secret,
                                    algorithms=[settings.token_algorithm])
         return decoded_token
-    except jwt.ExpiredSignatureError:
-        raise HTTPException(status_code=403, detail="Token expired")
-    except jwt.InvalidTokenError:
-        raise HTTPException(status_code=403, detail="Invalid token")
+    except jwt.ExpiredSignatureError as exc:
+        raise HTTPException(status_code=403, detail="Token expired") from exc
+    except jwt.InvalidTokenError as exc:
+        raise HTTPException(status_code=403, detail="Invalid token") from exc
 
 
 def encode_password(password):
-    hasher = CryptContext(schemes=["bcrypt"])
-    return hasher.hash(password)
+    return _hasher.hash(password)
 
 
 def verify_password(password, encoded_password):
-    hasher = CryptContext(schemes=["bcrypt"])
-    return hasher.verify(password, encoded_password)
+    return _hasher.verify(password, encoded_password)
