+++ app_v2/main.py
@@ -1,7 +1,7 @@
 from fastapi import FastAPI
 
 from config import settings
-from routers import categories, cities, users, queens
+from routers import categories, cities, users, queens, index
 
 app = FastAPI(title=settings.title,
               description=settings.description,
@@ -12,3 +12,4 @@ app.include_router(queens.router)
 app.include_router(categories.router)
 app.include_router(cities.router)
 app.include_router(users.router)
+app.include_router(index.router)
