+++ app_v2/requirements.txt
@@ -1,26 +1,18 @@
-anyio==3.6.2
-bcrypt==4.0.1
-certifi==2022.9.24
-charset-normalizer==2.1.1
-click==8.1.3
-colorama==0.4.6
-DateTime==4.7
-deta==1.1.0
-dnspython==2.2.1
-email-validator==1.3.0
-fastapi==0.86.0
-h11==0.14.0
-idna==3.4
-nose==1.3.7
-passlib==1.7.4
-pydantic==1.10.2
-PyJWT==2.6.0
-python-dotenv==0.21.0
-pytz==2022.6
-requests==2.28.1
-sniffio==1.3.0
-starlette==0.20.4
-typing_extensions==4.4.0
-urllib3==1.26.12
-uvicorn==0.19.0
-zope.interface==5.5.1
+anyio==3.6.2 ; python_version >= "3.10" and python_version < "4.0"
+bcrypt==4.0.1 ; python_version >= "3.10" and python_version < "4.0"
+click==8.1.3 ; python_version >= "3.10" and python_version < "4.0"
+colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
+deta==1.1.0 ; python_version >= "3.10" and python_version < "4.0"
+dnspython==2.2.1 ; python_version >= "3.10" and python_version < "4.0"
+email-validator==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
+fastapi==0.87.0 ; python_version >= "3.10" and python_version < "4.0"
+h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
+idna==3.4 ; python_version >= "3.10" and python_version < "4.0"
+passlib==1.7.4 ; python_version >= "3.10" and python_version < "4.0"
+pydantic==1.10.2 ; python_version >= "3.10" and python_version < "4.0"
+pyjwt==2.6.0 ; python_version >= "3.10" and python_version < "4.0"
+python-dotenv==0.21.0 ; python_version >= "3.10" and python_version < "4.0"
+sniffio==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
+starlette==0.21.0 ; python_version >= "3.10" and python_version < "4.0"
+typing-extensions==4.4.0 ; python_version >= "3.10" and python_version < "4.0"
+uvicorn==0.20.0 ; python_version >= "3.10" and python_version < "4.0"
