+++ app_v2/routers/index.py
@@ -1,0 +1,14 @@
+from fastapi import APIRouter
+from starlette.responses import RedirectResponse
+
+router = APIRouter(tags=["Index"])
+
+# region Index
+@router.get("/",
+            status_code=200,
+            include_in_schema=False)
+async def index_redirect():
+    response = RedirectResponse(url="/docs")
+    return response
+
+# endregion
