+++ app_v2/schemas/enums.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class StatusEnum(str, Enum):
+    Active = "Active"
+    Inactive = "Inactive"
+    Unknown = "Unknown"
